% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{median}
\alias{median}
\title{Calculate the Median of a Vector}
\usage{
median(vec)
}
\arguments{
\item{vec}{A numeric vector.}
}
\value{
The median value of the input vector.
}
\description{
This function calculates the median of a numeric vector using the \code{\link{quickselect}} algorithm.
}
\examples{
vec <- c(3, 1, 4, 2, 5)
median(vec)

}
