% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{quantile}
\alias{quantile}
\title{Calculate the Quantile of a Vector}
\usage{
quantile(vec, quantile)
}
\arguments{
\item{vec}{A numeric vector.}

\item{quantile}{The quantile value (between 0 and 1).}
}
\value{
The quantile value of the input vector.
}
\description{
This function calculates the quantile of a numeric vector using the \code{\link{quickselect}} algorithm.
}
\examples{
vec <- c(3, 1, 4, 2, 5)
quantile(vec, 0.25)

}
