% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{quickselect2}
\alias{quickselect2}
\title{Finds the k-th element in the sorted list efficiently}
\usage{
quickselect2(vec, k)
}
\arguments{
\item{vec}{A numeric vector.}

\item{k}{The desired rank of the element (1-based index).}
}
\value{
The k-th element in the sorted list.
}
\description{
This function finds the k-th element in the sorted list using the Quickselect algorithm.
}
\examples{
vec <- c(3, 1, 4, 2, 5)
quickselect(vec, 3)

}
